requirements_chatbot.txt
# Core LangChain and AI dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.0.0

# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Vector database and embeddings
weaviate-client>=3.25.0

# PDF processing
pypdf>=3.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Environment and utilities
python-dotenv>=0.19.0
